Fixes the plugin not respecting user settings if only one value was changed
If user sets the terminal to something else, like konsole but leaves the default execFlag, it won't be saved in config and this will result in empty execFlag when loading fom config. This fixes the issue by using default values, so if only one setting is changed, the other one will use default value.